Sayfalama ve Asenkron İşlemler Eklendi
IDisposable pattern'i ve ilgili kaynak dosyası kaldırıldı. IIsciView, IMusteriView ve ISatisView arayüzlerine sayfalama ve asenkron işlemler için yeni metodlar ve olaylar eklendi. IsciIdareetmeFormu ve MusteriIdareetmeFormu'na sayfalama UI elemanları eklendi. IsciPresenter ve MusteriPresenter'da sayfalama ve arama işlevselliği sağlandı. SatisPresenter'da işlemler asenkron hale getirildi ve yükleme göstergesi eklendi. LazyLoadComboBoxHelper'da zamanlayıcı daha performanslı bir alternatifle değiştirildi.

+++ AzAgroPOS.Teqdimat/Teqdimatcilar/BazaIdareetmePresenter.cs
@@ -61,7 +61,7 @@ public class BazaIdareetmePresenter
             var ölçüNetice = await _bazaManager.BazaOlcusunuGetirAsync();
             if (ölçüNetice.UgurluDur)
             {
-                _view.BazaOlcusunuGoster(ölçüNetice.Data);
+                _view.BazaOlcusunuGoster((double)ölçüNetice.Data);
             }
 
             // Son backup tarixini göstəririk
